+++ packages/stargate/src/queries/index.ts
@@ -7,6 +7,7 @@ export { QueryClient } from "./queryclient";
 export { AuthExtension, setupAuthExtension } from "./auth";
 export { BankExtension, setupBankExtension } from "./bank";
 export { DistributionExtension, setupDistributionExtension } from "./distribution";
+export { setupGovExtension, GovExtension, GovProposalId, GovParamsType } from "./gov";
 export { IbcExtension, setupIbcExtension } from "./ibc";
 export { setupStakingExtension, StakingExtension } from "./staking";
 export { createPagination, createProtobufRpcClient, ProtobufRpcClient } from "./utils";