+++ packages/stargate/src/queries/utils.ts
@@ -1,4 +1,5 @@
 import { Bech32 } from "@cosmjs/encoding";
+import { Uint64 } from "@cosmjs/math";
 import { PageRequest } from "cosmjs-types/cosmos/base/query/v1beta1/pagination";
 import Long from "long";
 
@@ -35,4 +36,13 @@ export function createProtobufRpcClient(base: QueryClient): ProtobufRpcClient {
       return base.queryUnverified(path, data);
     },
   };
+}
+
+/**
+ * Takes a uint64 value as string, number, Long or Uint64 and returns an unsigned Long instance
+ * of it.
+ */
+export function longify(value: string | number | Long | Uint64): Long {
+  const checkedValue = Uint64.fromString(value.toString());
+  return Long.fromBytesBE([...checkedValue.toBytesBigEndian()], true);
 }